auto_content_creator/requirements.txt
pyautogen==0.3.0
python-dotenv==1.0.0
black==24.8.0
flake8==7.1.1
requests==2.31.0
beautifulsoup4==4.12.2
scikit-learn==1.3.0
flaml[automl]==2.3.0
asyncpg==0.29.0
sse-starlette==2.1.3
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.4.2
anthropic